Change to new vendoring approach
Due to kustomize removing remote base support we've decided to go
with a vendoring approach for bases from the catalog.

This commit changes the build to:
 * support dashes in names (like e.g. nginx-ingress)
 * build all entries unless the build is triggered from a tag

Building all packges for builds not triggered by a tag is done to
ensure a change does not have any sideeffect that breaks the
build of another catalog entry.