Build test entry by default
Building all entries by default takes too long.

This change introduces a test entry under src/ which is now
used by default. Purpose of this is to have a way to test changes
to the pipeline and build automation.

To specifically build all entries branch names need to be prefixed
with `all-`.