Build module archives for Terraform
Terraform expects the module to be in the root of the archive. In
the interim while the website still shows the existing Kustomize
bases in the catalog we will build both an archive in the old
format (with a subdirectory inside the zip) and one for Terraform
with the files in the root of the archive.

Eventually, once the Terraform module work has been completed
we can stop building the old archives.