Document variant and additional_resources and set default_variant
default_variant is now set using a local specific to each service
because using a var to set the default leaves ambiguity for users
how to overwrite between the per workspace configuration or the
input variable directly.

Now the modules only expose the configuration and
configuration_base_key variables and the two module specific
and all the pass through Kustomization attributes are now handled
inside the per workspace configuration.