Remove namespaced ArgoCD variants, due to missing CRDs
Upstream removed the CRDs from the namespaced manifests, which
makes sense. But for Kubestack the focus is on cluster-wide
ArgoCD.

Users that prefer to install the namespaced variants,
can do so using the custom-manifests module. Use the module
to install the CRDs and then again once per namespace you
want a dedicated ArgoCD in.